Collect scout's output before parsing
Neovim 0.3 changed the way the output is sent to the job/terminal
process and now it sends a final EOF to indicate termination, but
because how scout works this is not enough to determine the final
selection sent by the program.

Now the plugin will collect scout's output and parse it at exit.
It will know what is the choice selected in scout because the program
signals the change of TTY with special chars sequence '?1049l'.